cheapamine & darksword

+++ Application/Dopamine/Jailbreak/DOEnvironmentManager.m
@@ -106,9 +106,9 @@ int reboot3(uint64_t flags, ...);
 {
     if (!gSystemInfo.jailbreakInfo.rootPath) {
         NSString *activePrebootPath = [self activePrebootPath];
-        
+
         NSString *randomizedJailbreakPath;
-        
+
         // First attempt at finding jailbreak root, look for Dopamine 2.x path
         for (NSString *subItem in [[NSFileManager defaultManager] contentsOfDirectoryAtPath:activePrebootPath error:nil]) {
             if (subItem.length == 15 && [subItem hasPrefix:@"dopamine-"]) {
@@ -116,7 +116,7 @@ int reboot3(uint64_t flags, ...);
                 break;
             }
         }
-        
+
         if (!randomizedJailbreakPath) {
             // Second attempt at finding jailbreak root, look for Dopamine 1.x path, but as other jailbreaks use it too, make sure it is Dopamine
             // Some other jailbreaks also commit the sin of creating .installed_dopamine, for these we try to filter them out by checking for their installed_ file
@@ -124,9 +124,9 @@ int reboot3(uint64_t flags, ...);
             for (NSString *subItem in [[NSFileManager defaultManager] contentsOfDirectoryAtPath:activePrebootPath error:nil]) {
                 if (subItem.length == 9 && [subItem hasPrefix:@"jb-"]) {
                     NSString *candidateLegacyPath = [activePrebootPath stringByAppendingPathComponent:subItem];
-                    
+
                     BOOL installedDopamine = [[NSFileManager defaultManager] fileExistsAtPath:[candidateLegacyPath stringByAppendingPathComponent:@"procursus/.installed_dopamine"]];
-                    
+
                     if (installedDopamine) {
                         // Hopefully all other jailbreaks that use jb-<UUID>?
                         // These checks exist because of dumb users (and jailbreak developers) creating .installed_dopamine on jailbreaks that are NOT dopamine...
@@ -136,7 +136,7 @@ int reboot3(uint64_t flags, ...);
                         if (installedNekoJB || installedPalera1n || installedDefinitelyNotAGoodName) {
                             continue;
                         }
-                        
+
                         randomizedJailbreakPath = candidateLegacyPath;
                         _bootstrapNeedsMigration = YES;
                         break;
@@ -144,7 +144,7 @@ int reboot3(uint64_t flags, ...);
                 }
             }
         }
-        
+
         if (randomizedJailbreakPath) {
             NSString *jailbreakRootPath = [randomizedJailbreakPath stringByAppendingPathComponent:@"procursus"];
             if ([[NSFileManager defaultManager] fileExistsAtPath:jailbreakRootPath]) {
@@ -161,7 +161,7 @@ int reboot3(uint64_t flags, ...);
     NSError *error = nil;
 
     [self locateJailbreakRoot];
-    
+
     if (!gSystemInfo.jailbreakInfo.rootPath || _bootstrapNeedsMigration) {
         [_bootstrapper ensurePrivatePrebootIsWritable];
 
@@ -175,11 +175,11 @@ int reboot3(uint64_t flags, ...);
             unichar randomCharacter = [characterSet characterAtIndex:randomIndex];
             [randomString appendFormat:@"%C", randomCharacter];
         }
-        
+
         NSString *randomJailbreakFolderName = [NSString stringWithFormat:@"dopamine-%@", randomString];
         NSString *randomizedJailbreakPath = [activePrebootPath stringByAppendingPathComponent:randomJailbreakFolderName];
         NSString *jailbreakRootPath = [randomizedJailbreakPath stringByAppendingPathComponent:@"procursus"];
-        
+
         if (_bootstrapNeedsMigration) {
             NSString *oldRandomizedJailbreakPath = [[NSString stringWithUTF8String:gSystemInfo.jailbreakInfo.rootPath] stringByDeletingLastPathComponent];
             [[NSFileManager defaultManager] moveItemAtPath:oldRandomizedJailbreakPath toPath:randomizedJailbreakPath error:&error];
@@ -189,12 +189,12 @@ int reboot3(uint64_t flags, ...);
                 [[NSFileManager defaultManager] createDirectoryAtPath:jailbreakRootPath withIntermediateDirectories:YES attributes:nil error:&error];
             }
         }
-        
+
         if (!error) {
             gSystemInfo.jailbreakInfo.rootPath = strdup(jailbreakRootPath.UTF8String);
         }
     }
-    
+
     return error;
 }
 
@@ -289,7 +289,7 @@ int reboot3(uint64_t flags, ...);
     if (ur == 0 && gr == 0) {
         rootBlock();
     }
-    
+
     if (gr == 0 && orgGroup != 0) setgid(orgGroup);
     if (ur == 0 && orgUser != 0) seteuid(orgUser);
 }
@@ -297,7 +297,7 @@ int reboot3(uint64_t flags, ...);
 - (int)runTrollStoreAction:(NSString *)action
 {
     if (![self isInstalledThroughTrollStore]) return -1;
-    
+
     uint32_t selfPathSize = PATH_MAX;
     char selfPath[selfPathSize];
     _NSGetExecutablePath(selfPath, &selfPathSize);
@@ -332,16 +332,41 @@ int reboot3(uint64_t flags, ...);
         __block int pid = 0;
         __block int r = 0;
         [self runUnsandboxed:^{
-            r = exec_cmd_suspended(&pid, JBROOT_PATH("/basebin/jbctl"), "reboot_userspace", NULL);
-            if (r == 0) {
-                // the original plan was to have the process continue outside of this block
-                // unfortunately sandbox blocks kill aswell, so it's a bit racy but works
+            r = exec_cmd_suspended(&pid, JBROOT_PATH("/usr/bin/killall"), "-9", "backboardd", NULL);
+            if (r == 0) kill(pid, SIGCONT);
 
-                // we assume we leave this unsandbox block before the userspace reboot starts
-                // to avoid leaking the label, this seems to work in practice
-                // and even if it doesn't work, leaking the label is no big deal
-                kill(pid, SIGCONT);
-            }
+            r = exec_cmd_suspended(&pid, JBROOT_PATH("/usr/bin/killall"), "-9", "mediaserverd", NULL);
+            if (r == 0) kill(pid, SIGCONT);
+
+            r = exec_cmd_suspended(&pid, JBROOT_PATH("/usr/bin/killall"), "-9", "installd", NULL);
+            if (r == 0) kill(pid, SIGCONT);
+
+            r = exec_cmd_suspended(&pid, JBROOT_PATH("/usr/bin/killall"), "-9", "commcenter", NULL);
+            if (r == 0) kill(pid, SIGCONT);
+
+            r = exec_cmd_suspended(&pid, JBROOT_PATH("/usr/bin/killall"), "-9", "identityservicesd", NULL);
+            if (r == 0) kill(pid, SIGCONT);
+
+            r = exec_cmd_suspended(&pid, JBROOT_PATH("/usr/bin/killall"), "-9", "sharingd", NULL);
+            if (r == 0) kill(pid, SIGCONT);
+
+            r = exec_cmd_suspended(&pid, JBROOT_PATH("/usr/bin/killall"), "-9", "SpringBoard", NULL);
+            if (r == 0) kill(pid, SIGCONT);
+
+            r = exec_cmd_suspended(&pid, JBROOT_PATH("/usr/bin/killall"), "-9", "lsd", NULL);
+            if (r == 0) kill(pid, SIGCONT);
+
+            r = exec_cmd_suspended(&pid, JBROOT_PATH("/usr/bin/killall"), "-9", "rapportd", NULL);
+            if (r == 0) kill(pid, SIGCONT);
+
+            r = exec_cmd_suspended(&pid, JBROOT_PATH("/usr/bin/killall"), "-9", "UserEventAgent", NULL);
+            if (r == 0) kill(pid, SIGCONT);
+
+            r = exec_cmd_suspended(&pid, JBROOT_PATH("/usr/bin/killall"), "-9", "wifid", NULL);
+            if (r == 0) kill(pid, SIGCONT);
+
+            r = exec_cmd_suspended(&pid, JBROOT_PATH("/usr/bin/killall"), "-9", "profiled", NULL);
+            if (r == 0) kill(pid, SIGCONT);
         }];
         if (r == 0) {
             cmd_wait_for_exit(pid);
@@ -482,7 +507,7 @@ int reboot3(uint64_t flags, ...);
     if (loaded) {
         [self setIDownloadEnabled:loaded needsUnsandbox:needsUnsandbox];
     }
-    
+
     void (^updateBlock)(void) = ^{
         if (loaded) {
             exec_cmd(JBROOT_PATH("/usr/bin/launchctl"), "load", JBROOT_PATH("/basebin/LaunchDaemons/com.opa334.Dopamine.idownloadd.plist"), NULL);
@@ -491,7 +516,7 @@ int reboot3(uint64_t flags, ...);
             exec_cmd(JBROOT_PATH("/usr/bin/launchctl"), "unload", JBROOT_PATH("/basebin/LaunchDaemons/com.opa334.Dopamine.idownloadd.plist"), NULL);
         }
     };
-    
+
     if (needsUnsandbox) {
         [self runAsRoot:^{
             [self runUnsandboxed:updateBlock];
@@ -500,7 +525,7 @@ int reboot3(uint64_t flags, ...);
     else {
         updateBlock();
     }
-    
+
     if (!loaded) {
         [self setIDownloadEnabled:loaded needsUnsandbox:needsUnsandbox];
     }
@@ -540,7 +565,7 @@ int reboot3(uint64_t flags, ...);
         [self runTrollStoreAction:@"hide-jailbreak"];
         return;
     }
-    
+
     void (^actionBlock)(void) = ^{
         BOOL alreadyHidden = [self isJailbreakHidden];
         if (hidden != alreadyHidden) {
@@ -564,7 +589,7 @@ int reboot3(uint64_t flags, ...);
             }
         }
     };
-    
+
     if ([self isJailbroken]) {
         [self runAsRoot:^{
             [self runUnsandboxed:actionBlock];
@@ -589,7 +614,7 @@ int reboot3(uint64_t flags, ...);
         if ([[NSFileManager defaultManager] fileExistsAtPath:kernelInApp]) {
             return kernelInApp;
         }
-        
+
         [[DOUIManager sharedInstance] sendLog:@"Downloading Kernel" debug:NO];
         NSString *kernelcachePath = [NSHomeDirectory() stringByAppendingPathComponent:@"Documents/kernelcache"];
         if (![[NSFileManager defaultManager] fileExistsAtPath:kernelcachePath]) {
@@ -602,7 +627,7 @@ int reboot3(uint64_t flags, ...);
 - (BOOL)isPACBypassRequired
 {
     if (![self isArm64e]) return NO;
-    
+
     if (@available(iOS 15.2, *)) {
         return NO;
     }
@@ -620,7 +645,7 @@ int reboot3(uint64_t flags, ...);
     //size_t cpuFamilySize = sizeof(cpuFamily);
     //sysctlbyname("hw.cpufamily", &cpuFamily, &cpuFamilySize, NULL, 0);
     //if (cpuFamily == CPUFAMILY_ARM_TYPHOON) return false; // A8X is unsupported for now (due to 4k page size)
-    
+
     DOExploitManager *exploitManager = [DOExploitManager sharedManager];
     if ([exploitManager availableExploitsForType:EXPLOIT_TYPE_KERNEL].count) {
         if (![self isPACBypassRequired] || [exploitManager availableExploitsForType:EXPLOIT_TYPE_PAC].count) {
@@ -629,7 +654,7 @@ int reboot3(uint64_t flags, ...);
             }
         }
     }
-    
+
     return false;
 }
